Move common functionality into utils file.
